// frontend/FreteEmbarcador-master/src/app/gfe/atualizacoes/cadastros/expedReceb/prazoEntrega/prazoEntrega.interface.ts
export interface PrazoEntrega {
  id?: number;
  filial?: string;
	codTabela: string;
	tipoTabela: string;
	idaVolta?: string;
	iniValidade: string;
	fimValidade: string;
	cidOrigem?: string;
	regOrigem?: string;
	remetente?: string;
	cidDest?: string;
	regDest?: string;
	destinatario?: string;
	tipoOper?: string;
	tipoVeiculo?: string;
	codGrupo?: string;
	transportador?: string;
	modal?: string;
	tipoPrazo?: string;
	prazo?: number;
	distEstimad?: number;
	prioridade?: number;
	classifFrete?: string;
	percMaxQuebra?: number;
	fretePeso?: number;
	aliqImposto?: number;
	solicCotacao?: string;
	email?: string;
	periodoSolic?: number;
	statusSolic?: string;
	dataUltAtu?: string;
	impostoInc?: string;
	tipoRef?: string;
	proxVigenc?: string;
	qtdDiasAnt?: number;
	qtdPrzResp?: number;
	diasLembret?: number;
	tipoFreq?: string;
	recorrencia?: string;
	modoSolicit?: string;
	sitSolicit?: string;
	sitEnvioWF?: string;
	unidPeso?: string;
	dataProxSol?: string;
	dataSolicEfet?: string;
	dataLimResp?: string;
	dataEnvioLemb?: string;
	dataLembrado?: string;
	dataResposta?: string;
	dataRespExp?: string;
	comunicar?: string;
	falhaEnvioWF?: string;
	userIncluir?: string;
	dataInclusao?: string;
	userAlterar?: string;
	dataAlteracao?: string;
  rotaOrigem?: string;
  rotaDestino?: string;
  descTpOper?: string;
  descTpVeic?: string;
  descGrupo?: string;
  descTransp?: string;
  descClass?: string;
}